Refactor error
